fix(api): apply route split review findings
X-Anneal-Run: cmthfpv9v305tmpxn4my0gpfy
X-Anneal-Step: 4: Apply review fixes
X-Anneal-Provider: codex
