+++ packages/api/src/routes/test-support.ts
@@ -1,12 +1,7 @@
 import "../test-workspace-root.js";
 import assert from "node:assert/strict";
-import { createHash } from "node:crypto";
 
-import {
-  RunnerKind,
-  RunnerPreference,
-  type PrismaClient,
-} from "@anneal/db";
+import { type PrismaClient } from "@anneal/db";
 
 import { createApp } from "../test-app.js";
 
@@ -35,122 +30,6 @@ export const lockedAgent = <T extends Record<string, unknown>>(agent: T | null):
   codexServiceTier: "DEFAULT",
   ...agent,
 }) : null;
-
-/* A merge-tail stop report: agent-authored text, attached to the task it
- * happened on, and nothing resumes on a reply. It is the shape the Inbox is
- * full of, and the old "attached to nothing" rule refused to close it. */
-export const stopReport = {
-  id: "message-1", status: "OPEN", from: "AGENT", kind: "TEXT", gateTaskId: null, replyToMessageId: null,
-};
-
-export const closeRequest = async (app: ReturnType<typeof createApp>, messageId: string): Promise<Response> =>
-  await app.request(`/inbox/messages/${messageId}/close`, {
-    method: "POST",
-    headers: { Authorization: "Bearer operator-unit-token", "Content-Type": "application/json" },
-    body: JSON.stringify({ requestId: `close-${messageId}` }),
-  });
-
-export const retryRequest = async (
-  assigneeAgent: {
-    id: string;
-    model: string;
-    runnerPreference: RunnerPreference;
-    foundationalPrompt: string;
-    rolePrompt: string;
-    archivedAt?: Date | null;
-    name?: string;
-  } | null,
-  templateStep: {
-    runner: RunnerKind | null;
-    stepIndex?: number;
-    outputKind?: string;
-    taskTemplate?: { name: string };
-  } | null = null,
-) => {
-  let created: Record<string, unknown> | undefined;
-  const currentTemplateStep = templateStep
-    ? { stepIndex: 1, outputKind: "result", taskTemplate: { name: "direct-engineer-workflow" }, ...templateStep }
-    : null;
-  const last = {
-    id: "run-1",
-    projectId: "project-1",
-    taskId: "task-1",
-    goalId: "goal-1",
-    agentId: "old-agent",
-    repoId: "repo-previous",
-    runNumber: 1,
-    status: "FAILED",
-    runner: RunnerKind.CLAUDE,
-    model: "old-model",
-    targetBranch: "main",
-    branch: "feature/retry",
-    promptHash: createHash("sha256").update("foundation\nrole\nRetry me\nUse current config").digest("hex"),
-    maxDurationMin: 90,
-    stallTimeoutMin: 7,
-    maxRunsPerTask: 4,
-    // Nothing granted, so the retry ceiling is the task's configured budget —
-    // which is what `maxRunsPerTask: 4` already was.
-    budgetGrants: 0,
-  };
-  const currentTask = {
-    id: "task-1",
-    projectId: "project-1",
-    name: "Retry me",
-    description: "Use current config",
-    assigneeType: "AGENT",
-    assigneeAgentId: assigneeAgent?.id ?? null,
-    repoId: "repo-current",
-    repo: null,
-    templateId: null,
-    templateStepId: currentTemplateStep ? "step-1" : null,
-    maxSessionsPerTask: 4,
-    maxDurationMin: 120,
-    stallTimeoutMin: 10,
-    opensPullRequest: true,
-    chainId: null,
-    chainIndex: null,
-    targetBranch: "main",
-    archivedAt: null,
-    dispatchAfterTaskId: null,
-    dispatchAfter: null,
-    assigneeAgent,
-    templateStep: currentTemplateStep,
-    runs: [last],
-  };
-  const database = {
-    $transaction: async (operation: (tx: unknown) => Promise<unknown>) => operation({
-      // Retry takes the shared task-row lock before it reads anything else.
-      $queryRaw: async () => [{ id: "task-1" }],
-      agent: { findUnique: async () => lockedAgent(assigneeAgent as Record<string, unknown> | null) },
-      task: {
-        findUniqueOrThrow: async () => ({ id: "task-1", status: "TODO", archivedAt: null }),
-        findUnique: async () => currentTask,
-        findMany: async () => [currentTask],
-        update: async () => ({}),
-      },
-      run: {
-        count: async () => 0,
-        groupBy: async () => [{
-          taskId: "task-1",
-          status: "FAILED",
-          _count: { _all: 1 },
-          _max: { budgetGrants: 0 },
-        }],
-        create: async ({ data }: { data: Record<string, unknown> }) => {
-          created = data;
-          return { id: "run-2", ...data };
-        },
-      },
-      agentRepoAccess: { count: async () => 1 },
-      taskActivity: { create: async () => ({}) },
-    }),
-  } as unknown as PrismaClient;
-  const response = await createApp(database).request("/tasks/task-1/retry", {
-    method: "POST",
-    headers: { Authorization: "Bearer operator-unit-token" },
-  });
-  return { response, created, last };
-};
 
 /* --------------------------------------------------- GET /tasks, projected */
 
@@ -203,29 +82,8 @@ export const getTasks = async (database: PrismaClient, query: string, headers: R
     headers: { Authorization: "Bearer operator-unit-token", ...headers },
   });
 
-export const taskDetailDatabase = (task: Record<string, unknown>): PrismaClient => ({
-  task: { findUnique: async () => task, findMany: async () => [task] },
-  run: { groupBy: async () => [] },
-  agentRepoAccess: { findMany: async () => [{ projectId: task.projectId, agentId: task.assigneeAgentId, repoId: task.repoId }] },
-  mergeRecoveryAttempt: { findFirst: async () => null },
-} as unknown as PrismaClient);
-
 export const untouchableDatabase = (): PrismaClient => new Proxy({}, {
   get(_target, property) {
     throw new Error(`the database must not be reached: ${String(property)}`);
   },
 }) as unknown as PrismaClient;
-
-export const onboardingBody = (overrides: Record<string, unknown> = {}): string => JSON.stringify({
-  project: { name: "My Project", slug: "my-project" },
-  repo: { name: "app", remoteUrl: "https://github.com/owner/name.git", defaultBranch: "main", mountPath: "repo" },
-  acknowledgedHostExecution: true,
-  ...overrides,
-});
-
-export const postOnboarding = async (database: PrismaClient, body: string): Promise<Response> =>
-  createApp(database).request("/onboarding", {
-    method: "POST",
-    headers: { Authorization: "Bearer operator-unit-token", "Content-Type": "application/json" },
-    body,
-  });
